Move Level, Subject and Tag into model/person
