+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -27,7 +27,7 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Note;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 
 /**
  * Edits the details of an existing person in the address book.
