+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -12,7 +12,7 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Note;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 
 /**
  * Contains utility methods for populating {@code AddressBook} with sample data.
