+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -40,7 +40,7 @@ import seedu.address.model.person.Address;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddCommandParserTest {
