+++ src/test/java/seedu/address/testutil/EditPersonDescriptorBuilder.java
@@ -11,7 +11,7 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Note;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 
 /**
  * A utility class to help with building EditPersonDescriptor objects.
