+++ src/test/java/seedu/address/testutil/PersonBuilder.java
@@ -9,7 +9,7 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Note;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
+import seedu.address.model.person.Tag;
 import seedu.address.model.util.SampleDataUtil;
 
 /**
